requirements.txt
streamlit
pandas
pytrends
plotly
requests
pillow
ntscraper
feedparser
